requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.1.0
numpy>=1.26.0
opencv-python>=4.8.0
ultralytics>=8.0.0
faiss-cpu>=1.7.0
insightface>=0.7.0
onnxruntime>=1.17.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0 